FIX: Avoid determining message store location when a stored message is edited
Revise range regexp again to also match an empty line above the boilerplate; this will be discarded by BufferPersist, anyway. We need a match in that case to properly replace a just-opened, otherwise empty commit message via CTRL-P or :MessageRecall. The only content that gets erroneously persisted is when the boilerplate is in line 1, since we cannot build a matching range starting in line 1 that contains only empty lines then.